February refactoring

+++ Assets/Scripts/Data/GameLevel/GameLevelInfo.cs
@@ -4,14 +4,14 @@
 [Serializable]
 public struct GameLevelInfo
 {
-    public float levelOneScore;
-    public float levelTwoScore;
-    public float levelThreeScore;
-    public float levelFourScore;
-    public float levelFiveScore;
-    public float levelSixScore;
-    public float levelSevenScore;
-    public float levelEightScore;
-    public float levelNineScore;
-    public float levelTenScore;
+    public int levelOneScore;
+    public int levelTwoScore;
+    public int levelThreeScore;
+    public int levelFourScore;
+    public int levelFiveScore;
+    public int levelSixScore;
+    public int levelSevenScore;
+    public int levelEightScore;
+    public int levelNineScore;
+    public int levelTenScore;
 }
